Remove third branch
It was not skipping the next machine correctly, leading to the process not finishing.

And actually, it was not needed, because the second branch skips one machine, and
if it's chosen on the next step, that's like skipping this and the next machine, and
that's what the third branch was trying to do.